DRY out logic to determine which jobs should be skipped
