+++ prow/pjutil/BUILD.bazel
@@ -31,6 +31,7 @@ go_library(
         "//vendor/github.com/satori/go.uuid:go_default_library",
         "//vendor/github.com/sirupsen/logrus:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
+        "//vendor/k8s.io/apimachinery/pkg/util/sets:go_default_library",
     ],
 )
