+++ prow/plugins/trigger/pull-request.go
@@ -226,11 +226,9 @@ func TrustedPullRequest(ghc githubClient, trigger plugins.Trigger, author, org, 
 func buildAll(c Client, pr *github.PullRequest, eventGUID string, elideSkippedContexts bool) error {
 	org, repo, number, branch := pr.Base.Repo.Owner.Login, pr.Base.Repo.Name, pr.Number, pr.Base.Ref
 	changes := config.NewGitHubDeferredChangedFilesProvider(c.GitHubClient, org, repo, number)
-	toTest, toSkipSuperset, err := pjutil.FilterPresubmits(pjutil.TestAllFilter(), changes, branch, c.Config.Presubmits[pr.Base.Repo.FullName], c.Logger)
+	toTest, toSkip, err := pjutil.FilterPresubmits(pjutil.TestAllFilter(), changes, branch, c.Config.Presubmits[pr.Base.Repo.FullName], c.Logger)
 	if err != nil {
 		return err
 	}
-
-	toSkip := determineSkippedPresubmits(toTest, toSkipSuperset, c.Logger)
 	return runAndSkipJobs(c, pr, toTest, toSkip, eventGUID, elideSkippedContexts)
 }
